+++ lib/main.dart
@@ -1,22 +1,17 @@
-import 'dart:typed_data';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:location/location.dart';
+import 'package:flutter_app/google.dart';
 import 'package:material_floating_search_bar/material_floating_search_bar.dart';
-import 'toilet.dart';
-import 'dart:ui' as ui;
 
-void main() => runApp(Test());
+void main() => runApp(AppWrapper());
 TabController _tabController;
 
-class Test extends StatefulWidget {
+class AppWrapper extends StatefulWidget {
   @override
-  _testAppState createState() => _testAppState();
+  _AppWrapperState createState() => _AppWrapperState();
 }
 
-class _testAppState extends State<Test> {
+class _AppWrapperState extends State<AppWrapper> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -31,89 +26,9 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> with TickerProviderStateMixin {
-  GoogleMapController mapController;
-  String mapStyling;
-
-  // Coordinates for DSV, Kista.
-  final LatLng _center = const LatLng(59.40672485297707, 17.94522607914621);
-  final int smallIconSize = 50;
-
-  Map<MarkerId, Marker> markers = <MarkerId, Marker>{};
-  Uint8List toiletIcon;
-
   initState() {
     super.initState();
     _tabController = TabController(length: 2, vsync: this);
-    rootBundle.loadString('map_style.txt').then((string) {
-      mapStyling = string;
-    });
-    loadIcons();
-  }
-
-  loadIcons() async {
-    toiletIcon = await getBytesFromAsset('assets/wc.png', smallIconSize);
-  }
-
-  Future<Uint8List> getBytesFromAsset(String path, int width) async {
-    ByteData data = await rootBundle.load(path);
-    ui.Codec codec = await ui.instantiateImageCodec(data.buffer.asUint8List(),
-        targetWidth: width);
-    ui.FrameInfo fi = await codec.getNextFrame();
-    return (await fi.image.toByteData(format: ui.ImageByteFormat.png))
-        .buffer
-        .asUint8List();
-  }
-
-  void _onMapCreated(GoogleMapController controller) {
-    if (mounted)
-      setState(() {
-        mapController = controller;
-        controller.setMapStyle(mapStyling);
-      });
-  }
-
-  void _currentLocation() async {
-    LocationData currentLocation;
-    var location = new Location();
-    try {
-      currentLocation = await location.getLocation();
-    } on Exception {
-      currentLocation = null;
-    }
-
-    mapController.animateCamera(CameraUpdate.newCameraPosition(
-      CameraPosition(
-        bearing: 0,
-        target: LatLng(currentLocation.latitude, currentLocation.longitude),
-        zoom: 17.0,
-      ),
-    ));
-  }
-
-  loadToilets() {
-    getToilets().then((toilets) {
-      for (Toilet t in toilets) {
-        String info = "Dritsatt: ${t.operational}\nAnpasad: ${t.adapted}";
-        addMarker(t.id.toString(), t.lat, t.long, "wc", info, toiletIcon);
-      }
-    });
-  }
-
-  addMarker(String id, double lat, double long, String type, String info,
-      Uint8List icon) {
-    MarkerId markerId = MarkerId(id + type);
-
-    final Marker marker = Marker(
-        icon: BitmapDescriptor.fromBytes(icon),
-        markerId: markerId,
-        position: LatLng(lat, long),
-        onTap: () {
-          print("marker tapped: ${markerId}");
-        });
-
-    setState(() {
-      markers[markerId] = marker;
-    });
   }
 
   Widget buildFloatingSearchBar() {
@@ -166,68 +81,67 @@ class _MyAppState extends State<MyApp> with TickerProviderStateMixin {
   }
 
   @override
-  Widget build(BuildContext context) => MaterialApp(
-      home: Scaffold(
-          resizeToAvoidBottomInset: false,
-          appBar: AppBar(
-            toolbarHeight: 48,
-            bottom: TabBar(
-              controller: _tabController,
-              tabs: <Widget>[
-                Tab(
-                  icon: Icon(Icons.map),
-                ),
-                Tab(
-                  icon: Icon(Icons.filter_alt_outlined),
-                ),
-              ],
-            ),
-          ),
-          body: Stack(
-            fit: StackFit.expand,
-            children: [
-              GoogleMap(
-                mapToolbarEnabled: false,
-                onMapCreated: _onMapCreated,
-                mapType: MapType.normal,
-                markers: Set<Marker>.of(markers.values),
-                zoomControlsEnabled: false,
-                myLocationEnabled: true,
-                myLocationButtonEnabled: false,
-                initialCameraPosition: CameraPosition(
-                  target: _center,
-                  zoom: 10,
-                ),
-              ),
-              buildFloatingSearchBar(),
-            ],
-          ),
-          floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
-          floatingActionButton: Column(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: <Widget>[
-              FloatingActionButton(
-                onPressed: _currentLocation,
-                child: Icon(Icons.search),
-              ),
-              FloatingActionButton(
-                onPressed: _currentLocation,
-                child: Icon(Icons.filter_alt_outlined),
-              ),
-              FloatingActionButton(
-                onPressed: _currentLocation,
-                child: Icon(Icons.location_on),
-              ),
-              FloatingActionButton(
-                onPressed: loadToilets,
-                child: Icon(Icons.airline_seat_legroom_extra),
-              ),
-              /**  FloatingActionButton(
-                      onPressed: goBack,
-                      child: Icon(Icons.home),
-                      ),**/
-            ],
-          ))
+  Widget build(BuildContext context) => MaterialApp(home: buildScaffold()
       //mainAxisAlignment: MainAxisAlignment.spaceBetween
       );
+  GoogleState _google = new GoogleState();
+  Scaffold buildScaffold() {
+    return Scaffold(
+        resizeToAvoidBottomInset: false,
+        appBar: buildAppBar(),
+        body: Stack(
+          fit: StackFit.expand,
+          children: [
+            _google.buildGoogleMap(),
+            buildFloatingSearchBar(),
+          ],
+        ),
+        floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
+        floatingActionButton: buildColumn());
+  }
+
+  AppBar buildAppBar() {
+    return AppBar(
+      toolbarHeight: 48,
+      bottom: TabBar(
+        controller: _tabController,
+        tabs: <Widget>[
+          Tab(
+            icon: Icon(Icons.map),
+          ),
+          Tab(
+            icon: Icon(Icons.filter_alt_outlined),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Column buildColumn() {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.end,
+      children: <Widget>[
+        FloatingActionButton(
+          onPressed: _google.currentLocation,
+          child: Icon(Icons.search),
+        ),
+        FloatingActionButton(
+          onPressed: _google.currentLocation,
+          child: Icon(Icons.filter_alt_outlined),
+        ),
+        FloatingActionButton(
+          onPressed: _google.currentLocation,
+          child: Icon(Icons.location_on),
+        ),
+        FloatingActionButton(
+          onPressed: _google.loadToilets,
+          child: Icon(Icons.wc_rounded),
+        ),
+        /**  FloatingActionButton(
+            onPressed: goBack,
+            child: Icon(Icons.home),
+            ),**/
+      ],
+    );
+  }
 }
